[FE] FIX: SearchItem 파일 분리 #202
